PostgreSQL'de daima üst geliştiriciyi tercih et

+++ lib/scripts/server/postgresql.sh
@@ -6,26 +6,22 @@ set -euo pipefail; [[ -z ${TRACE:-} ]] || set -x
 
 export DEBIAN_FRONTEND=noninteractive
 
-codename=$(lsb_release -sc)
+if [[ ! -f /etc/apt/sources.list.d/postgresql.list ]]; then
+	codename=$(lsb_release -sc)
 
-cat >/etc/apt/sources.list.d/postgresql.list <<-EOF
-	deb http://apt.postgresql.org/pub/repos/apt/ ${codename}-pgdg main
-EOF
-curl -fsSL https://www.postgresql.org/media/keys/ACCC4CF8.asc | apt-key add -
+	cat >/etc/apt/sources.list.d/postgresql.list <<-EOF
+		deb http://apt.postgresql.org/pub/repos/apt/ ${codename}-pgdg main
+	EOF
+	curl -fsSL https://www.postgresql.org/media/keys/ACCC4CF8.asc | apt-key add -
 
-cat >/etc/apt/preferences.d/postgresql.pref <<-EOF
-	Package: *
-	Pin: release o=apt.postgresql.org
-	Pin-Priority: 1000
-EOF
+	cat >/etc/apt/preferences.d/postgresql.pref <<-EOF
+		Package: *
+		Pin: release o=apt.postgresql.org
+		Pin-Priority: 1000
+	EOF
 
-apt-get -y update
-
-apt-get -y purge --auto-remove \
-	postgresql \
-	postgresql-client \
-	postgresql-contrib \
-	#
+	apt-get -y update
+fi
 
 apt-get -y install --no-install-recommends \
 	postgresql \
